letor: Sync testsuite with core

+++ xapian-letor/tests/harness/index_utils.cc
@@ -1,7 +1,7 @@
 /** @file
  * @brief utility functions for indexing testcase data
  */
-/* Copyright (C) 2005,2007,2013 Olly Betts
+/* Copyright (C) 2005,2007,2013,2026 Olly Betts
  *
  * This program is free software; you can redistribute it and/or modify
  * it under the terms of the GNU General Public License as published by
@@ -87,10 +87,10 @@ FileIndexer::index_to(Xapian::WritableDatabase & db)
             doc.add_value(i, para.substr(i, 1));
         }
         // Value 11 is useful for tests of sorting
-        doc.add_value(11, Xapian::sortable_serialise(para.size()));
+        doc.add_value(11, Xapian::sortable_serialise(double(para.size())));
 
         // Value 12 is useful for tests of collapsing
-        doc.add_value(12, Xapian::sortable_serialise(para.size() % 5));
+        doc.add_value(12, Xapian::sortable_serialise(double(para.size() % 5)));
 
         // Value 13 contains the first 3 letters of the paragraph
         doc.add_value(13, para.substr(0, 3));
